id 자바에서 Long 타입으로 수정

+++ Audio_mart/src/main/java/com/audio_mart/controller/HomeController.java
@@ -18,7 +18,7 @@ public class HomeController {
 	private MemberService memberService;
 	
 	private MemberDTO getMemberInfo(HttpSession session) {
-        Integer idx = (Integer) session.getAttribute("idx");
+        Long idx = (Long) session.getAttribute("idx");
         if (idx != null) {
             return memberService.findByIdx(idx);
         }
