+++ Audio_mart/src/main/java/com/audio_mart/mapper/ProductMapper.java
@@ -12,9 +12,9 @@ public interface ProductMapper {
 	
 	public boolean insertProduct(ProductDTO params);
 	public List<CategoryDTO> cateList();
-	public ProductDTO findByProductId(int productId);
+	public ProductDTO findByProductId(Long productId);
 	public boolean updateProduct(ProductDTO params);
-	public boolean deleteProduct(int productId);
+	public boolean deleteProduct(Long productId);
 	public List<ProductDTO> productList();
 
 }
